+++ DTCD-ExtensionRiskPrimitives/src/primitives/CalcNode/CalcNode_3.js
@@ -1,0 +1,91 @@
+import icon from './icon_3.svg';
+
+export default class CalcNode_3 {
+  static getPrimitiveInfo() {
+    return {
+      icon,
+      title: 'Расчетный',
+      name: 'Calc3',
+      groups: ['Рисковые примитивы 2'],
+    };
+  }
+
+  constructor(yFiles) {
+    this.yfiles = yFiles.default;
+  }
+
+  create() {
+    const {
+      SimpleNode,
+      ShapeNodeStyle,
+      Rect,
+      Font,
+      MarkupLabelStyle,
+      HorizontalTextAlignment,
+      TextWrapping,
+    } = this.yfiles;
+
+    const instance = new SimpleNode();
+
+    instance.layout = new Rect(0, 0, 294, 148);
+    instance.style = new ShapeNodeStyle({
+      shape: 'round-rectangle',
+      stroke: '4px #32ADE6',
+      fill: '#fff',
+    });
+
+    const customLabelStyle = new MarkupLabelStyle({
+      font: new Font('"Segoe UI", Arial', 12),
+      horizontalTextAlignment: HorizontalTextAlignment.LEFT,
+      wrapping: TextWrapping.WORD_ELLIPSIS,
+      insets: [20, 20],
+    });
+
+    const properties = {
+      type: { expression: `Расчетный`, type: 'expression' },
+      name: { expression: ``, type: 'expression' },
+      description: { expression: ``, type: 'expression' },
+      value: { expression: ``, type: 'expression' },
+    };
+
+    const initPorts = [
+      {
+        primitiveName: `inPort1`,
+        type: 'OUT',
+        portPosition: { x: 0.2, y: 1 },
+        properties: {
+          status: { expression: ``, type: 'expression' },
+        },
+      },
+      {
+        primitiveName: `inPort2`,
+        type: 'OUT',
+        portPosition: { x: 0.5, y: 1 },
+        properties: {
+          status: { expression: ``, type: 'expression' },
+        },
+      },
+      {
+        primitiveName: `inPort3`,
+        type: 'OUT',
+        portPosition: { x: 0.8, y: 1 },
+        properties: {
+          status: { expression: ``, type: 'expression' },
+        },
+      },
+      {
+        primitiveName: `outPort1`,
+        type: 'OUT',
+        portPosition: { x: 0.5, y: 0 },
+        properties: {
+          status: { expression: `value`, type: 'expression' },
+        },
+      }
+    ];
+
+    instance.tag = { customLabelStyle, properties, initPorts };
+
+    this.instance = instance;
+    return instance;
+  }
+}