+++ DTCD-ExtensionRiskPrimitives/src/primitives/FactorAnalysisNode/FactorAnalysisNode.js
@@ -1,11 +1,11 @@
 import icon from './icon.svg';
 
-export default class ExportNode {
+export default class FactorAnalysisNode {
   static getPrimitiveInfo() {
     return {
       icon,
-      title: 'Экспорт данных',
-      name: 'ExportNode',
+      title: 'Верхний блок факторный анализ',
+      name: 'FactorAnalysis',
       groups: ['Рисковые примитивы 2'],
     };
   }
@@ -30,7 +30,7 @@ export default class ExportNode {
     instance.layout = new Rect(0, 0, 294, 148);
     instance.style = new ShapeNodeStyle({
       shape: 'round-rectangle',
-      stroke: '4px #77dd77',
+      stroke: '4px #fde910',
       fill: '#fff',
     });
 
@@ -41,7 +41,27 @@ export default class ExportNode {
       insets: [20, 20],
     });
 
-    instance.tag = { customLabelStyle, properties: {}, initPorts: [] };
+    const properties = {
+      type: { expression: `Факторный анализ`, type: 'expression' },
+      name: { expression: ``, type: 'expression' },
+      description: { expression: ``, type: 'expression' },
+      value: { expression: ``, type: 'expression' },
+      management_action: { expression: ``, type: 'expression' },
+      external_influence: { expression: ``, type: 'expression' },
+    };
+
+    const initPorts = [
+      {
+        primitiveName: `inPort1`,
+        type: 'OUT',
+        portPosition: { x: 0.5, y: 1 },
+        properties: {
+          status: { expression: ``, type: 'expression' },
+        },
+      }
+    ];
+
+    instance.tag = { customLabelStyle, properties, initPorts };
 
     this.instance = instance;
     return instance;