+++ CHANGELOG.md
@@ -10,6 +10,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - new ExportNode primitive
+- new DataNode primitive
+- new CalcNode primitive
+- new GoalNode primitive
+- new ControlledRiskNode primitive
+- new UncontrolledRiskNode primitive
+- new RiskFactorNode primitive
+- new RiskNode primitive
+- new RiskMeasNode primitive
+- new MeasuresNode primitive
+- new KIRNode primitive
+- new AlertNode primitive
+- new PeriodSetupNode primitive
+- new FactorAnalysisNode primitive
+- new RiskRAppetiteNode primitive
 
 ## [0.2.0]